requirements.txt
# Tech-Stack NLP Analyzer (Germany Edition)
# Veri Madenciliği, NLP, Dashboard

requests>=2.28.0
pandas>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
matplotlib>=3.7.0

# AI yetenek çıkarma (opsiyonel)
openai>=1.0.0
